Convert embedding to float32 before returning
embeddings was bfloat16 (produced under torch.autocast(dtype=torch.bfloat16)), and NumPy has no bfloat16 dtype, so the final torch.cat(...).numpy() call raised TypeError: Got unsupported ScalarType BFloat16. Casting to float32 before moving off-GPU fixes it — only the small pooled embedding tensor is upcast, not the full model.

+++ mteb/models/model_implementations/audio_flamingo.py
@@ -125,7 +125,7 @@ class AudioFlamingoWrapper(AbsEncoder):
                 batch_indices = torch.arange(hidden.size(0), device=hidden.device)
                 embeddings = hidden[batch_indices, last_idx]
 
-                all_embeddings.append(embeddings.cpu().detach())
+                all_embeddings.append(embeddings.float().cpu().detach())
 
         return torch.cat(all_embeddings, dim=0).numpy()
